Modifying the SampleAlgoRunner to be able to run an IAlgo class in a predefined namespace
The goal is the  SampleAlgoRunner to be able to run an algo from blob by id.
There could be a single algo comming from a blob.
The algo should be placed in namespace "Lykke.AlgoStore.CSharp.Algo.Implemention.ExecutableClass"
The algo class could have any name, but should implement the IAlgo interface and be the only such class in the namespace.

By default there will be no such class in the namespace. During a build of specific algo the class will be generated.

The SampleAlgorunner is now aware of the above. It will try to run the single class meeting the criteria above. If more than one algo class is meeting the criteria it will notify by throwing exception. If none - by default it will behave as it was before - run the DummyAlgo class